+++ example/example.g.dart
@@ -82,23 +82,23 @@ Item _$ItemFromJson(Map<String, dynamic> json) => new Item()
 // **************************************************************************
 
 final _$glossaryDataJsonLiteral = {
-  "glossary": {
-    "title": "example glossary",
-    "GlossDiv": {
-      "title": "S",
-      "GlossList": {
-        "GlossEntry": {
-          "ID": "SGML",
-          "SortAs": "SGML",
-          "GlossTerm": "Standard Generalized Markup Language",
-          "Acronym": "SGML",
-          "Abbrev": "ISO 8879:1986",
-          "GlossDef": {
-            "para":
-                "A meta-markup language, used to create markup languages such as DocBook.",
-            "GlossSeeAlso": ["GML", "XML"]
+  'glossary': {
+    'title': 'example glossary',
+    'GlossDiv': {
+      'title': 'S',
+      'GlossList': {
+        'GlossEntry': {
+          'ID': 'SGML',
+          'SortAs': 'SGML',
+          'GlossTerm': 'Standard Generalized Markup Language',
+          'Acronym': 'SGML',
+          'Abbrev': 'ISO 8879:1986',
+          'GlossDef': {
+            'para':
+                'A meta-markup language, used to create markup languages such as DocBook.',
+            'GlossSeeAlso': ['GML', 'XML']
           },
-          "GlossSee": "markup"
+          'GlossSee': 'markup'
         }
       }
     }