+++ test/test_files/json_literal.g.dart
@@ -1,0 +1,62 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of json_serializable.example;
+
+// **************************************************************************
+// Generator: JsonLiteralGenerator
+// **************************************************************************
+
+final _$dataJsonLiteral = [
+  'simple string',
+  "'string with single quotes'",
+  '"string with double quotes"',
+  '\'With singles and \"doubles\"\'',
+  r'dollar $igns',
+  r"'single quotes and dollor $ig$'",
+  r"${'nice!'}",
+  '""hello""',
+  r'""$double quotes and dollar signs""',
+  '\$scary with \'single quotes\' and triple-doubles \"\"\"oh no!',
+  null,
+  true,
+  false,
+  5,
+  5.5351,
+  -5.5,
+  {},
+  {
+    'null': null,
+    'int': 42,
+    'double': 42.0,
+    'string': 'string',
+    'list': [],
+    'bool': true
+  }
+];
+const _$asConstJsonLiteral = const [
+  'simple string',
+  "'string with single quotes'",
+  '"string with double quotes"',
+  '\'With singles and \"doubles\"\'',
+  r'dollar $igns',
+  r"'single quotes and dollor $ig$'",
+  r"${'nice!'}",
+  '""hello""',
+  r'""$double quotes and dollar signs""',
+  '\$scary with \'single quotes\' and triple-doubles \"\"\"oh no!',
+  null,
+  true,
+  false,
+  5,
+  5.5351,
+  -5.5,
+  const {},
+  const {
+    'null': null,
+    'int': 42,
+    'double': 42.0,
+    'string': 'string',
+    'list': const [],
+    'bool': true
+  }
+];